requirements.txt
discord.py
aiohttp
python-decouple
psutil
asyncpg
